# lab/openapi.yaml
openapi: 3.1.0
info:
  title: oat lab — articles
  version: "1.0"
  description: >
    A backend the oat repo owns. Iterate the handler, this document, and oat
    independently. Two tenants (org_alpha / org_beta) authenticate with API keys.
paths:
  /v1/auth/token:
    post:
      operationId: auth.token
      summary: Exchange an API key for a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses:
        "200":
          description: Credential
          content:
            application/json:
              schema:
                type: object
                required: [access_token, org_id]
                properties:
                  access_token: { type: string }
                  org_id: { type: string }
        "401":
          description: Unknown key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/invites/{token}/accept:
    post:
      operationId: invite.accept
      summary: Accept an invite
      x-entity: { name: article, action: action, identity: id }
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted: { type: boolean }
        "404":
          description: Unknown invite
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
...
  schemas:
...
    Error:
      type: object
      required: [error_key, message]
      properties:
        error_key: { type: string }
        message: { type: string }